Sending level as uint32_t instead of int

+++ Projects/CoX/Common/GameData/GameDataStore.cpp
@@ -255,15 +255,15 @@ bool GameDataStore::read_runtime_data(const QString &directory_path)
     return true;
 }
 
-uint32_t GameDataStore::expForLevel(int lev) const
+uint32_t GameDataStore::expForLevel(uint32_t lev) const
 {
-    assert(lev>0 && lev<(int)m_experience_and_debt_per_level.m_ExperienceRequired.size());
+    assert(lev>0 && lev<m_experience_and_debt_per_level.m_ExperienceRequired.size());
     return m_experience_and_debt_per_level.m_ExperienceRequired.at(lev - 1);
 }
 
-uint32_t GameDataStore::expDebtForLevel(int lev) const
+uint32_t GameDataStore::expDebtForLevel(uint32_t lev) const
 {
-    assert(lev>0 && lev<(int)m_experience_and_debt_per_level.m_DefeatPenalty.size());
+    assert(lev>0 && lev<m_experience_and_debt_per_level.m_DefeatPenalty.size());
     return m_experience_and_debt_per_level.m_DefeatPenalty.at(lev - 1);
 }
